Adapt to current Master (2)
