+++ src/test/java/ch/uzh/ifi/hase/soprafs21/rest/mapper/DTOMapperActivityTest.java
@@ -1,6 +1,7 @@
 package ch.uzh.ifi.hase.soprafs21.rest.mapper;
 
 import ch.uzh.ifi.hase.soprafs21.constant.ActivityCategory;
+import ch.uzh.ifi.hase.soprafs21.constant.SwipeStatus;
 import ch.uzh.ifi.hase.soprafs21.entities.Activity;
 import ch.uzh.ifi.hase.soprafs21.entities.ActivityPreset;
 import ch.uzh.ifi.hase.soprafs21.entities.User;
@@ -9,6 +10,7 @@ import ch.uzh.ifi.hase.soprafs21.rest.dto.activityDTO.ActivityGetDTO;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
@@ -56,4 +58,34 @@ class DTOMapperActivityTest {
         assertEquals(activity.getUserSwipeStatusList().get(0).getUser(),activityGetDTO.getUserSwipeStatusList().get(0).getUser());
         assertEquals(activity.getUserSwipeStatusList().get(0).getSwipeStatus(),activityGetDTO.getUserSwipeStatusList().get(0).getSwipeStatus());
     }
+
+
+    @Test
+    public void test_Activity_to_ActivityGetDTO() {
+        // create user
+        User user = new User();
+        user.setId(1L);
+        user.setEmail("firstname@lastname");
+        user.setPassword("verySafePassword");
+        user.setToken("token123");
+
+        // set user swipe status
+        UserSwipeStatus userSwipeStatus = new UserSwipeStatus(user, SwipeStatus.INITIAL);
+        List<UserSwipeStatus> userSwipeStatusList = Collections.singletonList(userSwipeStatus);
+
+        // set activity preset
+        ActivityPreset activityPreset = new ActivityPreset("Test Activity", ActivityCategory.MUSIC);
+
+        // combine user swipe status and activity preset to activity
+        Activity activity = new Activity(activityPreset, userSwipeStatusList);
+        List<Activity> activityList = Collections.singletonList(activity);
+
+        // map
+        List <ActivityGetDTO> activityGetDTOList = DTOMapperActivity.INSTANCE.convertEntityListToActivityGetDTOList(activityList);
+
+        // check content
+        assertEquals(activityList.get(0).getActivityPreset(), activityGetDTOList.get(0).getActivityPreset());
+        assertEquals(activityList.get(0).getUserSwipeStatusList(), activityGetDTOList.get(0).getUserSwipeStatusList());
+    }
+
 }
