--- src/test/java/ch/uzh/ifi/hase/soprafs21/rest/mapper/DTOMapperTest.java
@@ -1,98 +1,0 @@
-package ch.uzh.ifi.hase.soprafs21.rest.mapper;
-
-import ch.uzh.ifi.hase.soprafs21.constant.ActivityCategory;
-import ch.uzh.ifi.hase.soprafs21.constant.Gender;
-import ch.uzh.ifi.hase.soprafs21.constant.SwipeStatus;
-import ch.uzh.ifi.hase.soprafs21.entities.Activity;
-import ch.uzh.ifi.hase.soprafs21.entities.ActivityPreset;
-import ch.uzh.ifi.hase.soprafs21.entities.User;
-import ch.uzh.ifi.hase.soprafs21.entities.UserSwipeStatus;
-import ch.uzh.ifi.hase.soprafs21.rest.dto.activityDTO.ActivityGetDTO;
-import ch.uzh.ifi.hase.soprafs21.rest.dto.userDTO.UserGetDTO;
-import ch.uzh.ifi.hase.soprafs21.rest.dto.userDTO.UserPostDTO;
-import org.junit.jupiter.api.Test;
-
-import java.time.LocalDateTime;
-import java.time.temporal.ChronoUnit;
-import java.util.Collections;
-import java.util.Date;
-import java.util.List;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-/**
- * DTOMapperTest
- * Tests if the mapping between the internal and the external/API representation works.
- */
-public class DTOMapperTest {
-    @Test
-    public void testCreateUser_fromUserPostDTO_toUser_success() {
-        // create UserPostDTO
-        UserPostDTO userPostDTO = new UserPostDTO();
-        userPostDTO.setEmail("name@uzh.ch");
-        userPostDTO.setPassword("blabla");
-        userPostDTO.setName("Tester");
-
-        // MAP -> Create user
-        User user = DTOMapperUser.INSTANCE.convertUserPostDTOtoEntity(userPostDTO);
-
-        // check content
-        assertEquals(userPostDTO.getEmail(), user.getEmail());
-        assertEquals(userPostDTO.getPassword(), user.getPassword());
-        assertEquals(userPostDTO.getName(), user.getName());
-    }
-
-    @Test
-    public void testGetUser_fromUser_toUserGetDTO_success() {
-        // create User
-        User user = new User();
-        user.setEmail("firstname@lastname");
-        user.setLastSeen(LocalDateTime.now().truncatedTo(ChronoUnit.MINUTES));
-        user.setName("test");
-        user.setGender(Gender.MALE);
-        user.setBio("asdf");
-        user.setToken("1");
-
-        // MAP -> Create UserGetDTO
-        UserGetDTO userGetDTO = DTOMapperUser.INSTANCE.convertEntityToUserGetDTO(user);
-
-        // check content
-        assertEquals(user.getId(), userGetDTO.getId());
-        assertEquals(user.getEmail(), userGetDTO.getEmail());
-        assertEquals(user.getLastSeen(), userGetDTO.getLastSeen());
-        assertEquals(user.getName(),userGetDTO.getName());
-        assertEquals(user.getGender(),userGetDTO.getGender());
-        assertEquals(user.getBio(),userGetDTO.getBio());
-        assertEquals(user.getToken(),userGetDTO.getToken());
-    }
-
-    @Test
-    public void test_Activity_to_ActivityGetDTO() {
-        // create user
-        User user = new User();
-        user.setId(1L);
-        user.setEmail("firstname@lastname");
-        user.setPassword("verySafePassword");
-        user.setToken("token123");
-
-        // set user swipe status
-        UserSwipeStatus userSwipeStatus = new UserSwipeStatus(user, SwipeStatus.INITIAL);
-        List<UserSwipeStatus> userSwipeStatusList = Collections.singletonList(userSwipeStatus);
-
-        // set activity preset
-        ActivityPreset activityPreset = new ActivityPreset("Test Activity", ActivityCategory.MUSIC);
-
-        // combine user swipe status and activity preset to activity
-        Activity activity = new Activity(activityPreset, userSwipeStatusList);
-        List<Activity> activityList = Collections.singletonList(activity);
-
-        // map
-        List <ActivityGetDTO> activityGetDTOList = DTOMapperActivity.INSTANCE.convertEntityListToActivityGetDTOList(activityList);
-
-        // check content
-        assertEquals(activityList.get(0).getActivityPreset(), activityGetDTOList.get(0).getActivityPreset());
-        assertEquals(activityList.get(0).getUserSwipeStatusList(), activityGetDTOList.get(0).getUserSwipeStatusList());
-    }
-
-
-}
